[FAB-6107] unit test for bulk loading & a bug fix
As discussed in https://gerrit.hyperledger.org/r/#/c/13149/, need a unit
test to ensure cache is loaded correctly during validation phase when
using private data.

This CR introduces GetCachedVersion() in BulkOptimizable interface so
that we can test the cache loading logic. GetCachedVersion() retrieves
key's version present in the cache (and if not present, it returns false)
whereas GetVersion() retrieves version from state db if required key is
not available in the cache.

This CR fixes a bug in LoadCommittedVersions() implemented in
privacyenabledstate.DB.

This CR also renames ClearCommittedVersions() to ClearCachedVersions()
as it is more relevant.

Change-Id: Ib04ea26db025c82c4e34be48a7d875a5363ef3b7
Signed-off-by: senthil <cendhu@gmail.com>