[FAB-6186] MSP cache should not use RWMutex
The MSP cache has RW locks and uses RLock when doing cache lookup,
and WLock when doing cache updates.

This may result in the an error if done concurrently while the cache
has more than 1 item and only cache lookups are performed,
because the cache lookup operation shifts and moves items in the
internal list of the cache (that specifies eviction order,
since its an LRU cache).

The test was changed and would fail wit a null pointer panic
if the cache.go file would be reverted back.

Change-Id: I97f8b8cce0ee21090da0b036814da52acbf3be60
Signed-off-by: yacovm <yacovm@il.ibm.com>