[FAB-6195] Update one-line description of commit msg
This CR updates the documentation to include the Jira issue ID in
the online commit message for readability. It also states that issue
IDs in commit messages will be automatically linked to the corresponding
entry in Jira.

Change-Id: Ic5a79137db10f1ba6fcc434e0632a9448e2cffd9
Signed-off-by: Dileban Karunamoorthy <dileban@gmail.com>

[FAB-6195] Use [FAB-XXXX] in commit message example

The convention right now is to specify the Jira issue ID with
square brackets.

Change-Id: Ic5a79137db10f1ba6fcc434e0632a9448e2cffd9
Signed-off-by: Dileban Karunamoorthy <dileban@gmail.com>

[FAB-6195] Fix typo in description following example

A silly typo, should be 'of' instead of 'if'

Change-Id: Ic5a79137db10f1ba6fcc434e0632a9448e2cffd9
Signed-off-by: Dileban Karunamoorthy <dileban@gmail.com>