[FAB-5591] Revise timeout when launching chaincode
Previous timer covers only the startup phase - and does not
include building the chaincode image. This is not a user-friendly
definition.

This CR will ensure that the timer starts before trying to build
the image, which would make more sense from the user's perspective
to represent the time from beginning to end of the instantiation
request message.

Refactored launch related code to aid with UT and added UT.

Change-Id: Ie106642c60b8e1b6aadff384c40dd3308681e855
Signed-off-by: Boliang Chen <cblsjtu@gmail.com>
Signed-off-by: Srinivasan Muralidharan <srinivasan.muralidharan99@gmail.com>