[FAB-6294] Fix stale reference to policy manager
The policy manager used to be stateful, and it was safe to retain a
reference to it.  However, this added considerable complexity and
introduced the possibility for subtle races so it was turned into an
immutable structure.

The orderer signature filter was not updated to pull a fresh reference
on each invocation, but instead still retains only a reference to the
original policy manager.  This CR updates the signature filter to fetch
a fresh reference upon each invocation.

Change-Id: Ibaf8a18cc992feaf737047c87ce8973450aa1d19
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>