[FAB-6364] Make keepalive policy less restrictive
Currently the keepalive interval used for
client connections (e.g. peer to orderer)
is set to 5 min.  This setting is also
used to determine the the threshold for
how often the server will allow pings from
clients.  5 min seemed like a good default,
but in practice 1 min seems to be a better
option

Change-Id: I2f2262dfda8320d1f1731d7fd9eacf497e8cef58
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>