[FAB-6431] Pvt Transient Store to use received height
Transient store persists private data on authorized
peers while transactions with private data are in-flight.
One of the keys for transient store records was
endorsementBlkHt, this work item will change it to
blockHeight, as it will be better to store transient data
based on 'received at' height rather than endorsement
height, when a peer receives private data via gossip.
The height will be the basis for block-to-live purging
in subsequent changesets.

Also, this change updates comments to indicate that
TxPvtReadWriteSet includes write information but not
read information (the read information is passed in the
hashed rwset rather than in the private set). A
subsequent changeset will rename TxPvtReadWriteSet
to TxPvtWriteSet throughout the code.

Change-Id: I52df4702ad47ca17f2a745faa4c278f78e350d2f
Signed-off-by: David Enyeart <enyeart@us.ibm.com>