[FAB-6485] Add capabilities hook for MSP version
The MSP framework needs to know which version of MSP semantics to use
for principal evaluation etc.  This needs to be determined based on the
capabilities defined by the channel.  This CR adds a new method to the
capabilities interface for the channel which returns an MSP version.

It also modifies the MSPConfigHandler to accept this version, but for
the time being, does nothing with it.  There is a TODO left in the code
for where this flag should be used by the new MSP code which consumes
it.

Change-Id: Ic23be11b6814ae830ebd113442c365f7dd58de9b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>