[FAB-6512] Change gossip leader election default
The sampleconfig/core.yaml file defaults to disabling leader election
and instead setting the peer to be a leader.

This is the wrong default, because enabling leader election fixes a
number of potential errors, including peers permanently disconnectoin
from the ordering service during a network partition, and allows a peer
to join a channel with a genesis block which does not contain the peer's
org definition.

Change-Id: I1f6de1d5e7a0275003450756e6303e70643a04ca
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>