[FAB-6517] Make unit tests faster
make unit-tests currently uses the gocov
command and this seems to make things run
much slower than simply running go test.

gocov was used to generate coverage output
compatible with the Cobertura format as
Jenkins uses a Cobertura plugin.

It turns out that it is possible to combine
coverage reports from multiple runs of
go test and then convert the output to
a Cobertura compatible format.

Change-Id: I63ee27a950cac48d6d27b25b78842246c37a2065
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>