[FAB-6399] Benign panic in config update computation
This CR adds a check for the length of a config item's path and
triggers a panic if it is zero. The decision to still trigger
a panic stems from the fact that if the item's path is zero, it
means something unexpected occurred in the validation logic and
we should not proceed. The user who originally reported this bug
likely hit a config issue that is no longer possible after the
fix for FAB-5606.

Change-Id: I4052174c04e0810714823c3ca8a533ab2f261804
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>