[FAB-6660] Add filtered actions array to filt. block
This CR adds an array of filtered actions (each with its own chaincode
event, if set) to the filtered transaction proto. This will keep the
filtered block consistent with the block while also preparing for any
future additions to the transaction action that should also be included
in the filtered transaction's action.

Change-Id: I5b5c320ccf0ed6bd20c3983d487492f189c16264
Signed-off-by: Will Lahti <wtlahti@us.ibm.com>