[FAB-6751] Remove encshim interface
encshim exposed a confusing interface to permit chaincode-level encryption
since it wrapped GetState and PutState but it didn't wrap GetStateByRange
(which had to be handled directly with EncrypterEntity). This change set fixes
that, removing encshim and amending the example to show how encryption can be
performed only by using the entitites interface.

Change-Id: Ice24f9655baccd7d574b2878f689edfb0af86f67
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>