[FAB-6839] configtx to directly utilize cb.Config
The configtx package has changed considerably over time, and no longer
needs to be initialized with an envelope message.  The only places which
actually instantiate the configtx manager are creating an artificial
Envelope message and mashaling the config inside it, purely to satisfy
the constructor.  This CR modifies the constructor to take the
ultimately necessary type.

Change-Id: I3e4909e45118694b309f733a5179efb7d9a235fd
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>