[FAB-6877] Add resourcesconfig into peer singleton
The peer singleton presently utilizes a channelconfig.BundleSource for
handling channel configuration.  To support the new resources tree,
which is a superset of the channel configuration, this needs to be
converted to a resourcesconfig.BundleSource.  This CR does just that.
It does not actually implement parsing of the resources tree from the
genesis block, which will be added in a future CR.

Change-Id: I4fcdb35af1b44ddd9db1911215763ec91f8f9d73
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>