[FAB-7044] Refactor gRPC server config code
When creating a new core/comm#GRPCServer instance,
the only options which are publicly exposed are
those related to security.  All other config
such as keepalive and message size settings are
hard-coded and/or set internally.

This change creates a new top-level ServerConfig
struct which will be used in future changesets
to hold various configuration options.  Initially
it holds SecureOptions (which used to be
SecureServerConfig).

Change-Id: Ie6cd8c74ea9579504eb2bcba384498fede386cd7
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>