[FAB-7095] configurable pulling of cc base images
Currently, when chaincode is instantiated, the pulling of baseos
is explicitly set to false. This commit adds a new configuration
option under the docker section which allows one to force the
pulling of baseos. By default, it is set to false which maintains
the current behaviour.

This change may help address situations where a base image is updated
(eg: to patch vulnerabilities) and a peer cluster is using a generic
moving tag for baseos such as "latest".

Change-Id: I39af030ce26dcd3282ec4e762f8ed217ea6dfe9b
Signed-off-by: gbolo <george.bolo@gmail.com>