[FAB-6987]Fix single_tx_client panic error
Likley due to the fact that it has not
really been maintained, the
single_tx_client panicked because
the proper message structures were
not being send to the orderer.

Once this was fixed, it led to
a type mismatch error due to the
way the deliver channel is set up.

Rather than muck around with this, I
simply got the client working such that
it gets the latest block from the orderer
and then also gets the block for the tx it
submits as well.

Change-Id: Ie05ac79a2d940e2b30b1d8935bb2feeea527518b
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>