[FAB-6234] configtxgen emit resource defaults
This changeset at its core, simply enables embedding a default resource
tree into the output of channel creation transactions which are created
with v1.1 application capabilities.

Unit tests have been added. Additionally, the new channel creation tx
behavior can be confirmed by issuing:

 configtxgen -profile SampleSingleMSPChannelV1_1 \
  -outputCreateChannelTx foo.tx -inspectChannelCreateTx foo.tx

and observing the resources tree in the JSON output.

Because the utility function for creating the channel creation
transaction is used widely throughout the tests, and the function
parameters needed to change significantly, there is a fair amount of
churn in the tests, but none in production code.

Change-Id: Id719a1ae3f309a0a189f566dc0fef409564b519e
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>