[FAB-7239] Update Kafka client
Updated to github.com/Shopify/sarama@6a8d89d

Not claiming support for Kafka 0.11 or 1.0 with this changeset, but
preliminary testing (e.g. running the bddtests) shows that we work
with those versions of Kafka.

Declaration of Kafka 1.0 support is left as a separate issue pending
some further testing (which this patch enables).

Change-Id: Ib5f72bfc8077645b22eb489f51fe4749617225f9
Signed-off-by: Luis Sanchez <sanchezl@us.ibm.com>