[FAB-7322] configtxgen set default res mod_policy
In v1.0, chaincode instantiation was always controlled by a policy which
required a single signature.  By default, we wish for chaincode
instantiation to require a quorum of administrators to sign off, but
doing so may break existing workflows.

This CR adds a field to configtx.yaml which specifies the default
mod_policy to use throughout the resources tree.  If set to a single
signature policy like /Channel/Application/Writers, then the old workflow
of a single signer is still possible.  If set to a multi-sig policy like
/Channel/Application/Admins, then the user's workflow will need to be
changed to accomodate the additional signing requirements.

Change-Id: I021801f1540ee5ee25d110682852564650421dbd
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>