[FAB-7306] delivery-client OSN black-listing fine-tune
When the peer receives a bad status (i.e UNAVAILABLE) from
an ordering service node that it connects to, it disconnects
from it and tries another ordering service node.

FAB-5006 introduced an optimization for v1.1 that makes the peer
temporarily black-list ordering service nodes that send a bad status
in order for the peer to prioritize selection of other OSNs,
instead of connecting to them with the same probability.

The black-listing works via calling Disconnect() with a 'true' parameter.

However - this might be problematic in cases where the OSNs are
behind a LB/NAT where only 1 endpoint is actually known to the peer,
but in practice - several endpoints exist behind.

This would result in the peer always disabling immediately all OSNs
because it only knows 1 endpoint.

This change set overcomes this by avoiding black-listing an endpoint
if it is the last one remaining.

Change-Id: I59ead8af88eb9acff3d3c64227528b18b990fb00
Signed-off-by: yacovm <yacovm@il.ibm.com>