[FAB-7156] Fix for an error ignored in commit path
A specific error (when during validation trying to get the
current version of an item from the db) is ignored. This
results in a non-deterministinc behavior because the error
could happen only on a few peers (e.g., couchdb is inaccessible from the peer).

This CR propagates the error up which would result in returing the block commit
operation with an error instead of marking the transaction as invalid on the
peer facing intermittent issue with db.

Change-Id: I4022196cd8c8200891f52f6b7a69c8a062c793cd
Signed-off-by: manish <manish.sethi@gmail.com>