[FAB-7884] InitTLSForPeer fail on missing caCert config
Prior to the change set, the method InitTLSForPeer falls
back on an empty root CA cert pool if the config option isn't specified.

It doesn't make sense to do that, since we don't have any way of trusting
a remote node in such a case and thus establishing a TLS connection.

Change-Id: I7b0595995d879f3d44b1fd616e80f44aefb473d4
Signed-off-by: yacovm <yacovm@il.ibm.com>