[FAB-7525] Strong ciphers for TLS
While Fabric only supports TLS 1.2,
it currently uses the default set of
ciphers provided by Go.  This change
sets the default ciphers to a very
strong set (AESGCM).

While the code is structured to
allow these to be configured in
the future, this change does not
allow the ciphers to be configured.

Change-Id: I6bb68454adacadac63bef3ce89c25e4c687e482b
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>