[FAB-7535] Signature policy SatisfiedBy set
In order for the discovery service to tell which peers are
needed to endorse a certain chaincode, it needs to query
the policy to obtain all combinations of principals,
such that each combination can satisfy the policy
given enough peer signatures have been collected.

This change set creates an InquireablePolicy which is an
enhanced version of a Policy and currently only signature policies
can be used to create it (more to be added in the future).

The InquireablePolicy has a method SatisfiedBy() []PrincipalSet
which computes a slice of PrincipalSet.
A PrincipalSet is a slice of MSP Principals (with repetitions),
such that if enough signatures are collected from peers that
satisfy the principal, the entire policy can be satisifed.

In other words, the InquireablePolicy computes ways of satisfying
the endorsement policy.

Change-Id: I9e8315096cd18478854d219eab0c9a91eb7a05d6
Signed-off-by: yacovm <yacovm@il.ibm.com>