[FAB-6651] check instantiation policy upon upgrade
When a chaincode is being upgrade, there might be 2 instantiation policies
worth checking: that of the current chaincode and that of the upgraded
chaincode. The existing code checks that the request complies with the
existing instantiation policy (i.e. the one on the ledger). The new code
also checks whether the request complies with the new instantiation policy
(i.e. the one on the rwset). This code change affects the behaviour of the
committer and is therefore enclosed in a capability block.

Change-Id: I6dbdba3ee13a48f165e07c4cee802042a76ab92d
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>