[FAB-7554] Include Envelopes in gossip membership API
The discovery service needs to provide clients with envelopes of
peers so that the clients could (optionally) verify the signatures
themselves if they desire.
The discovery.NetworkMember already has an envelope from previous
change sets, and we need to populate the field when we return
the envelopes in the discovery and channel modules.

Change-Id: If582553c4e19d3403c3f8728b3a6fb5a3f8e27ef
Signed-off-by: yacovm <yacovm@il.ibm.com>