Use of invalid path
Hi,

First, love the docs, one of the only projects on Earth that actually really document how to generate offline viewable docs, bravo.

For the local build, it states "starting from the main fabric" directory implies your in the directory; the enumerated steps show otherwise as if your above it.  Updated to assume you are in the directory per the previous statement.

Happy New Year!