[FAB-7573] cleanup chaincode support
Allowing concurrent access to invokes and init in 1.0
allowed all the 0.6 separation of Query, Invoke and Init
to be vastly simplified (in 0.6 Queries were allowed
concurrent access but not others).

This also implied that the FSM can be removed completely
which was done in FAB-5424. This task completes the work
with the following

change names of functions to not have "FSM" flavor (they
now have "handler" flavor) in core/chaincode/handler.go
    . chaincode and peer calls driven by function handlers
    . comment and sundry cleanup
    . This was tested as follows

peer and shim chaincode UT
    . destructive testing by killing peer and chaincode at
      different points of the lifecycle
    . running against a java chaincode to prove the two sides
      are independent and shims can be worked upon independent
      of peer chaincode support side

Change-Id: I16e59eb2c77e6cc8f0bcfafccb0ae7b220c7317e
Signed-off-by: Srinivasan Muralidharan <srinivasan.muralidharan99@gmail.com>