[FAB-7580] simplify READY state
The chaincode support code in the peer sends a READY message to
the chaincode after receiving notification of registration. The
code path can be shortened (basically avoid a channel notification
followed by call to send READY) where the READY is sent immediately
following receipt of register from chaincode.

This code and codepath reduction further simplifies the chaincode
cleanup.

This was tested with
   . UT
   . ccchecker
   . go CC devmode
   . java Chaincode

Change-Id: I7cfb85d4c05e9073bb628d74c84013ce29ad77bc
Signed-off-by: Srinivasan Muralidharan <srinivasan.muralidharan99@gmail.com>