[FAB-7710] Only allow SCC plugins with build flag
It is currently possibly to extend the peer with
custom system chaincode(s) deployed as Go plugins.
Extending the peer requires the all members of a
network have peers with the plugins deployed.

In order to ensure that this is a conscientious
decision, this capability will not be enabled
by default and instead will require the peer to
be built using a build tag named "pluginsenabled"

Change-Id: Iec68a22c4a038b48433b21cd679962c39b6187eb
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>