[FAB-7757] CouchDB unit tests fail for CouchDB 2.1.1
Index definitions in CouchDB 2.1.1 allow for a Partial Filter Selector
in the index definition.   When an index definition is returned by
CouchDB, it now includes this as part of the JSON response.

The unit test needs to verify that the correct field and sort is in
the JSON response and ignore the rest of the index definition.

Change-Id: Iaeefd1dfee6b1b918b061925c74280637458c842
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>