[FAB-7676] Ignore read-only collections preimage pull
Currently, when the coordinator parses the block, it doesn't distinguish
between collections that have no private writes and collections that do.

This results in gossip attempting to retrieve from the transient store, or from
other peers - preimages for read-only collection RWSets, which don't exist anywhere
because they're not persisted in the transient store, and as a result - the attempt
times out.

This change set makes the logic skip collections that are read-only.

Change-Id: If41d2e0ebe093d567c480d17d9b7e26263659498
Signed-off-by: yacovm <yacovm@il.ibm.com>