[FAB-8019] CouchDB Retry count is misleading
The CouchDB retry currently assumes the first pass through is
actually retry number 1.

It also does not currently allow a setting of zero retries.
This can be confusing to the end user and will prevent some potential
use cases that may need to wrap the request handler with zero retry logic.

Change-Id: I74a5628f19fd6d73a24091bad66f955046e28190
Signed-off-by: Chris Elder <chris.elder@us.ibm.com>