updated the doc with correct CLI default exit time
The default CLI exit time is 10 sec as per the configuration while following this document. & I found this as a big issue and very hard to notice and can be a big impediment for anyone who is following this doc. 
If you want to update this information by yourself, you can do it in your language. I've added the important details already as an edit to this document.