[FAB-8065] Properly override logger backend in test
The test TestHaltChainProcessing attempts to restore the
log backend by overriding the logger, but the function that
creates the logger, caches previous logger instances by the package name.

Even though we can override the logger instance with a new package name,
it is probably more correct to overwrite the backend itself.

Change-Id: Iba4f6cb6586f41cd3b7a019896a3dac0d8c58d95
Signed-off-by: yacovm <yacovm@il.ibm.com>